pyproject.toml
[project]
name="space_flight_control_system"
version="0.0.1"
license = {text = "MIT License"}
requires-python = "~= 3.10"
dependencies=[
    "django~=5.0.3",
    "djangorestframework~=3.15.1",
    "drf-spectacular~=0.27.1",
    "apscheduler~=3.10.4",
    "django-rest-enumfield~=0.2.0",
]

[build-system]
requires = ["setuptools>=68.2.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []